LayoutTests/http/tests/notifications/legacy/request-no-callback-expected.txt

This test checks that requesting permission without a callback works. This supports legacy API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Requesting permission with default permission
PASS window.webkitNotifications.checkPermission() is 1
PASS webkitNotifications.requestPermission() did not throw exception.
Requesting permission with non-default permission
PASS window.webkitNotifications.checkPermission() is 2
PASS webkitNotifications.requestPermission() did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
